Improved Helpers Package: Refactor timedate to datetime
